app/requirements.txt
pydantic==2.7.1
requests==2.31.0
python-dotenv==1.0.1
fastapi
sqlalchemy
sqlmodel